+++ lib/rspec/rails/example/routing_example_group.rb
@@ -1,4 +1,5 @@
 require "action_dispatch/testing/assertions/routing"
+require "rspec/rails/example/routing_example_group/hyper_shortcut"
 
 module RSpec::Rails
   module RoutingExampleGroup
@@ -19,6 +20,8 @@ module RSpec::Rails
     end
 
     included do
+      self.instance_eval{extend HyperShortcut::AdditionalMethods}
+
       metadata[:type] = :routing
 
       before do