+++ lib/rspec/rails/example/routing_example_group/hyper_shortcut.rb
@@ -1,0 +1,6 @@
+require "rspec/rails/example/routing_example_group/hyper_shortcut/behavior"
+require "rspec/rails/example/routing_example_group/hyper_shortcut/matcher_placeholder"
+require "rspec/rails/example/routing_example_group/hyper_shortcut/request_pair"
+require "rspec/rails/example/routing_example_group/hyper_shortcut/shortcut_elements"
+require "rspec/rails/example/routing_example_group/hyper_shortcut/subject_placeholder"
+require "rspec/rails/example/routing_example_group/hyper_shortcut/additional_methods"