+++ lib/rspec/rails/example/routing_example_group/hyper_shortcut/matcher_placeholder.rb
@@ -1,0 +1,12 @@
+module RSpec::Rails::HyperShortcut
+  class MatcherPlaceholder
+    def initialize(name,args)
+      @name = name
+      @args = args
+    end
+
+    def build_matcher_in(example)
+      example.send(@name,*@args)
+    end
+  end
+end