+++ lib/rspec/rails/example/routing_example_group/hyper_shortcut/request_pair.rb
@@ -1,0 +1,16 @@
+module RSpec::Rails::HyperShortcut
+  class RequestPair
+    def initialize(method,path)
+      @method = method
+      @path = path
+    end
+
+    def to_hash
+      {@method => @path}
+    end
+
+    def to_s
+      "#{@method.to_s.upcase} #{@path}"
+    end
+  end
+end