+++ spec/rspec/rails/example/routing_example_group_spec.rb
@@ -1,5 +1,4 @@
 require "spec_helper"
-
 module RSpec::Rails
   describe RoutingExampleGroup do
     it { should be_included_in_files_in('./spec/routing/') }